pyproject.toml
[project]
name = "persona-switcher-mcp"
version = "1.0.0"
description = "MCP server for dynamic AI persona management with file-based storage"
authors = [{name = "User"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "fastmcp>=2.0.0",
    "python-frontmatter>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
auto-reload = [
    "watchdog>=3.0.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
